Separate path creation logic
